+++ blockchain/indexers/error.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2021 The Decred developers
+// Copyright (c) 2021-2022 The Decred developers
 // Use of this source code is governed by an ISC
 // license that can be found in the LICENSE file.
 
@@ -13,9 +13,6 @@ type ErrorKind string
 const (
 	// ErrUnsupportedAddressType indicates an unsupported address type.
 	ErrUnsupportedAddressType = ErrorKind("ErrUnsupportedAddressType")
-
-	// ErrInvalidSpendConsumerType indicates an invalid spend consumer type.
-	ErrInvalidSpendConsumerType = ErrorKind("ErrInvalidSpendConsumerType")
 
 	// ErrConnectBlock indicates an error indexing a connected block.
 	ErrConnectBlock = ErrorKind("ErrConnectBlock")
