+++ rpctest/node.go
@@ -110,8 +110,6 @@ func (n *nodeConfig) arguments() []string {
 	args = append(args, fmt.Sprintf("--rpckey=%s", n.keyFile))
 	// --txindex
 	args = append(args, "--txindex")
-	// --addrindex
-	args = append(args, "--addrindex")
 	if n.dataDir != "" {
 		// --datadir
 		args = append(args, fmt.Sprintf("--datadir=%s", n.dataDir))
